fix lint and some more refactoring of hooks
