+++ src/canvas/components/WidgetContainer.tsx
@@ -1,4 +1,4 @@
-import { memo, RefObject, useCallback, useRef } from "react";
+import { RefObject, memo, useCallback, useRef } from "react";
 import Draggable, { DraggableData, DraggableEvent } from "react-draggable";
 import { ErrorBoundary } from "react-error-boundary";
 import ErrorDisplay from "../components/ErrorDisplay";
